fix build problem with spaces in paths

+++ src/parameter_study.cpp
@@ -699,8 +699,8 @@ double get_n_target(const OdeSolution &sol, const ControlParameters &cpar)
     const Parameters* par = Parameters::get_parameters(cpar.mechanism);
     if (par == nullptr) return 0.0;
     if (cpar.target_specie == par->invalid_index) return 0.0;
-    if (sol.x.front().size() != 4 + par->num_species) return 0.0;
-    if (sol.x.back().size() != 4 + par->num_species) return 0.0;
+    if (sol.x.front().size() != size_t(4 + par->num_species)) return 0.0;
+    if (sol.x.back().size() != size_t(4 + par->num_species)) return 0.0;
 
     const double R_last = 100.0 * sol.x.back()[0];  // [cm]
     const double V_last = 4.0 / 3.0 * std::numbers::pi * std::pow(R_last, 3); // [cm^3]
